Emoji
Added CSS for emoji as an exercise to figure out how I want to go
about this. I'm thinking a lot of how options are going to work is
going to revolve around manipulating the main CSS as opposed to the
userstyles.org method of adding blocks of code to it. The emoji
option does both right now, so we'll see how that works.