More rice.
There's a bug in the options where the riced checkboxes are
unresponsive, but the labels work fine. This greatly concernes me.

I think the labels are forcing two click events to be fired when
the .rice div is clicked? I'll look into it later. I am quite
satisfied with how far I've gotten as of now.