Fix edited object saving without user confirmation.
Turns out assigning an object to a variable references it
rather than copying it. Oops.

On the positive side, this bug would only occur if you cancelled an
edit and then saved an edit on the same object or object's parent,
and I caught it before it was reported.

Go me.