Faster Catalog Links code
Ours is based on @ihavenoface's old catalog code, not @MayhemYDG's,
so we don't particularly benefit from advancements in his code, per
se, without rechecking and validating the quality of our own. This
happens to be a case where we were stuck in an inefficient paradigm
and never revisited.

We now appear much closer to Mayhem's code, but are still mostly
absent from being merged into the rest of the Header code. Whether
we decide to merge or not should probably be discussed at some
point.