Turns out I never finished my Thread Title magic...
because @MayhemYDG had already done it and done it better. So I
thought mine was working when it was just his... and then
realized when I didn't disconnect it and the index was getting a
title from a thread that it wasn't working and his was.

So let's just use his code.