Better $.rmAll
I verified an issue with the previous benchmark and even wrote a
more efficient method based off el.cloneNode(false);, which I would
use if cloneNode could inherit eventListeners

Speaking of events. there is still no interface for gather all
eventListeners on a given element (or its children). Apparently the
issue is that the w3c doesn't think there is a legitimate use case
for it. Heh.