Re-think the Quote Backlink implementation.
Before this rewrite, containers and backlinks were created even if the relevant
post did not exist in the page. If the post would never end up existing, it
resulted in a waste of resources. QuoteBacklink's cache also prevented these
containers and backlinks from being garbage collected.
We now only store a map of fullIDs. This feature should be faster since less
elements gets allocated, although the difference is probably negligible.