Disallow custom HTML in file info format.
I doubt anyone is using it, and if someone manages to inject a
malicious script into the page, this could be used to install a
script permanently to be executed each time 4chan is visited.

I'll mention it in the changelog, and if anyone complains, I can
add it back with some security checks.