Remove warosu.org (hopefully temporarily) (#32)
warosu.org is having problems with its hosting provider, and has been offline for the past few days (2-ish). The Bibliotheca Anonoma is currently in contact with the admin of the site, and we are working to find a new host and/or create a copy of all the data.