Better handling of content blockers that throw exceptions during AJAX requests. #789, #910
Adblock Plus and NoScript throw the error on `r.send()` now.

Expand try-catch block to handle it, and add check that it's
really a content-blocker error.

Remove the notification; this should be handled the same as a
connection error since that's what other content blockers
(e.g. uBlock Origin) will simulate. Instead log connection errors
that don't return a status code to the console.