Add box to reverse index, appends -rev if checked
When box element is toggled, invoke a revised
sort option that appends "-rev" to the sort
dropdown value when box is checked. If this value
changes while box is checked, the new sort is
still reversed.

If "-rev" is already given by previously saved
sort or custom board navigation option, pick the
correct dropdown value and check box
automatically.

Wrap previous sorting implementation, determining
if "-rev" is tacked on and doing the reverse work
at the end.