Added filename substring method to Sauce
Usage %sname[start,len] or %sname[len] - returns a substring of length
len from position start. Allows for start to have negative values,
wrapping around the end of the filename. extension and delimiting period
are excluded from the filename string but are re-appended at the end.
"0 or a" puzzles me.